+++ src/index.ts
@@ -14,3 +14,4 @@ export * from "./form/file-picker";
 export * from "./form/file-field";
 export * from "./form/date-picker";
 export * from "./form/date-field";
+export * from "./form/time-picker";